+++ Wiki/ReleaseNotes.md
@@ -2,9 +2,19 @@
 ====
 ## Version 6
 
+### 6.0.2
+
+- Update references
+- Fixed nuspec
+- Added ability to remove metrics
+- Added ability to parametrize concurrent checkpoin tracker flush strategy.
+
+### 6.0.1
+
+- Update references.
+
 ### 6.0.0
 
-- Added AggregateVersion on standard Readmodel (non atomic)
 - Tested and used with .NET core projects.
 - Removed Metrics.NET in favor of [App Metrics](https://www.app-metrics.io/)
 - Support usage of Rebus MongoDb transport